requirements.txt
blinker==1.9.0
click==8.1.8
filelock==3.18.0
Flask==3.1.1
fsspec==2025.7.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
pandas==2.3.1
pillow==11.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sympy==1.14.0
torch==2.8.0
torchvision==0.23.0
typing_extensions==4.14.1
tzdata==2025.2
Werkzeug==3.1.3
zipp==3.23.0
